+++ NEWS.md
@@ -1,3 +1,18 @@
+# robust2sls 0.2.1
+
+## Major changes
+
+* addition of Impulse Indicator Saturation (IIS) as initial estimator for outlier detection
+(from ivgets package)
+
+## Minor changes
+
+* code coverage has been extended to 100%
+* update formulae for FODR under normality (estimation not required)
+* all 2SLS estimation commands originally from AER package now use the ivreg package
+* fixed \eqn or \deqn or with empty second argument, as suggested by CRAN maintainers
+
+
 # robust2sls 0.2.0
 
 ## Major changes